simulator/simulator/requirements.txt
bokeh==3.1.1
click==8.1.3
cloudpickle==2.2.1
contourpy==1.0.7
dask==2023.4.1
distributed==2023.4.1
et-xmlfile==1.1.0
fsspec==2023.5.0
HeapDict==1.0.1
importlib-metadata==6.6.0
jdcal==1.4.1
Jinja2==3.1.2
locket==1.0.0
MarkupSafe==2.1.2
msgpack==1.0.5
numpy==1.22.0
openpyxl==3.0.4
packaging==23.1
pandas==2.0.1
partd==1.4.0
Pillow==9.5.0
psutil==5.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
python-i18n==0.3.9
pytz==2023.3
PyYAML==6.0
scipy==1.10.1
six==1.16.0
sortedcontainers==2.4.0
tblib==1.7.0
toolz==0.12.0
tornado==6.3.1
tzdata==2023.3
urllib3==2.0.2
xlrd==1.2.0
xyzservices==2023.2.0
zict==3.0.0
zipp==3.15.0

